+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Show the wallpaper behind the preview when configuring a widget
 
+### Fixed
+
+* Fix "empty folder" being displayed before items actually load
+
 ## [v1.43.0] - 2026-06-20
 [v1.43.0]: https://github.com/fibelatti/photo-widget/releases/tag/v1.43.0